Changed power column in user creatures table to biginteger

+++ database/migrations/2022_01_09_074027_add_power_column_to_user_creature_table.php
@@ -14,7 +14,7 @@ class AddPowerColumnToUserCreatureTable extends Migration
     public function up()
     {
         Schema::table('bio10_user_creatures', function (Blueprint $table) {
-            $table->integer('power')->after('species_id');
+            $table->bigInteger('power')->after('species_id');
         });
     }
